requirements.txt
Package                    Version
-------------------------- -----------
altair                     5.4.1
attrs                      24.2.0
blinker                    1.8.2
cachetools                 5.5.0
certifi                    2024.8.30
cffi                       1.17.1
chardet                    5.2.0
charset-normalizer         3.3.2
chromedriver-autoinstaller 0.6.4
ci-info                    0.3.0
click                      8.1.7
colorama                   0.4.6
configobj                  5.0.9
configparser               7.1.0
cryptography               43.0.1
etelemetry                 0.3.1
exceptiongroup             1.2.2
Faker                      25.2.0
filelock                   3.16.1
fitz                       0.0.1.dev2
Flask                      3.0.3
gitdb                      4.0.11
GitPython                  3.1.43
h11                        0.14.0
httplib2                   0.22.0
idna                       3.10
importlib_resources        6.4.5
isodate                    0.6.1
itsdangerous               2.2.0
Jinja2                     3.1.4
joblib                     1.4.2
jsonschema                 4.23.0
jsonschema-specifications  2023.12.1
looseversion               1.3.0
lxml                       5.3.0
markdown-it-py             3.0.0
MarkupSafe                 2.1.5
mdurl                      0.1.2
narwhals                   1.8.4
networkx                   3.3
nibabel                    5.3.0
nipype                     1.8.6
numpy                      1.26.4
opencv-python              4.10.0.84
outcome                    1.3.0.post0
packaging                  24.1
pandas                     2.2.2
pathlib                    1.0.1
pdf2image                  1.17.0
pdfminer                   20191125
pdfminer.six               20231228
pdfplumber                 0.11.4
pillow                     10.4.0
pip                        24.2
protobuf                   5.28.2
prov                       2.0.1
pyarrow                    17.0.0
pycparser                  2.22
pycryptodome               3.21.0
pydeck                     0.9.1
pydot                      3.0.2
Pygments                   2.18.0
PyMuPDF                    1.24.11
pyparsing                  3.1.4
PyPDF2                     3.0.1
pypdfium2                  4.30.0
PySocks                    1.7.1
pytesseract                0.3.13
python-dateutil            2.9.0.post0
pytz                       2024.1
pyxnat                     1.6.2
rdflib                     6.3.2
referencing                0.35.1
requests                   2.32.3
rich                       13.8.1
rpds-py                    0.20.0
scikit-learn               1.5.0
scipy                      1.13.1
selenium                   4.25.0
setuptools                 63.2.0
simplejson                 3.19.3
six                        1.16.0
smmap                      5.0.1
sniffio                    1.3.1
sortedcontainers           2.4.0
streamlit                  1.38.0
tenacity                   8.5.0
threadpoolctl              3.5.0
toml                       0.10.2
tornado                    6.4.1
traits                     6.3.2
trio                       0.26.2
trio-websocket             0.11.1
typing_extensions          4.12.2
tzdata                     2024.1
urllib3                    2.2.3
watchdog                   4.0.2
websocket-client           1.8.0
Werkzeug                   3.0.3
wsproto                    1.2.0